change code as per comments
